backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# LangGraph and LangChain
langgraph==0.0.26
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24

# OpenAI
openai==1.12.0

# Supabase
supabase==2.3.4
python-jose[cryptography]==3.3.0

# Tavily for web search
tavily-python==0.3.3

# Utilities
httpx==0.26.0
tiktoken==0.6.0
python-dotenv==1.0.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

